250203_Main UI

+++ Assets/02. Scripts/Player/PlayerDetailsSO.cs
@@ -30,5 +30,5 @@ public class PlayerDetailsSO : IdentifiedObject
     [Space(10)]
     [Header("Unlock")]
     public bool isUnlock; // 해금 여부
-    public float unlockCost; // 해금 비용
+    public int unlockCost; // 해금 비용
 }
